feat: use tmp x2 & cleanup
