+++ loader/palera1nLoader/UI/Views/ActionsVC.swift
@@ -157,7 +157,7 @@ class ActionsVC: UIViewController, UITableViewDelegate, UITableViewDataSource {
             case 1:
                 #if !targetEnvironment(simulator)
                 let alert = UIAlertController.warning(title: local("ACTION_HIDEJB"), message: local("HIDE_NOTICE"), destructiveBtnTitle: local("PROCEED"), destructiveHandler: {
-                    if fileExists("/var/mobile/Library/palera1n/helper") {
+                    if fileExists("/tmp/palera1n/helper") {
                         if (!envInfo.isRootful) && fileExists("/var/jb") {
                             binpack.rm("/var/jb")
                         }
